Issue #2543: extracted locale check from KillSwitchLayout
